+++ src/views/home_page.html
@@ -36,7 +36,7 @@
         rgba(45, 42, 70, 0.55)
         ),
         /* bottom, image */
-        url("../img/maryland.png");
+        url("../static/maryland.png");
 
         /* Full height */
         height: 100%;
