+++ src/views/search_page.html
@@ -89,7 +89,7 @@
         rgba(45, 42, 70, 0.55)
         ),
         /* bottom, image */
-        url("../img/maryland.png");
+        url("../static/maryland.png");
 
         /* Full height */
         height: 100%;
